pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sre_defender_env"
version = "0.1.0"
description = "SRE Defender environment for OpenEnv — Meta x Scaler Hackathon"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=15.0.1",
    "httpx>=0.28.1",
    "openai>=2.7.2",
    "requests>=2.31.0",
    "aiohttp==3.13.5",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m sre_defender_env.server.app
server = "sre_defender_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sre_defender_env", "sre_defender_env.server"]
package-dir = { "sre_defender_env" = ".", "sre_defender_env.server" = "server" }